+++ frontend/src/App.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 
-import BaseRouter from "./routes";
+import BaseRouter from "./Routes";
 import { checkTokenExpiration } from "./store/actions/auth";
 
 import "react-toastify/dist/ReactToastify.min.css";
