+++ frontend/src/containers/notes/Create.js
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+
+import Layout from "../../components/Layout";
+
+class NoteCreate extends Component {
+  render = () => {
+    return (
+      <Layout>
+        <div className="notes create">Note create</div>
+      </Layout>
+    );
+  };
+}
+
+export default NoteCreate;
