+++ frontend/src/containers/notes/List.js
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+
+import Layout from "../../components/Layout";
+
+class NotesList extends Component {
+  render = () => {
+    return (
+      <Layout>
+        <div className="notes list">Notes list</div>
+      </Layout>
+    );
+  };
+}
+
+export default NotesList;
